refactor: Projects have been moved in folders by meaning.
