+++ .github/workflows/dotnetcore.yml
@@ -20,6 +20,8 @@ jobs:
       uses: actions/setup-dotnet@v1
       with:
         dotnet-version: 6.0.x
+    - name: Change source code folder
+      run: cd src
     - name: Install dependencies
       run: dotnet restore
     - name: Build
