--- AbstractBuilder/Internal/CtorConstants.cs
@@ -1,12 +1,0 @@
-namespace AbstractBuilder.Internal
-{
-    internal static class CtorConstants
-    {
-        internal const string MethodName = ".ctor";
-
-        internal const System.Reflection.BindingFlags BindingFlags =
-            System.Reflection.BindingFlags.Instance |
-            System.Reflection.BindingFlags.Public |
-            System.Reflection.BindingFlags.NonPublic;
-    }
-}
