--- UnitTests/Examples/Builders/PointBuilder.cs
@@ -1,13 +1,0 @@
-namespace AbstractBuilder.Examples.Builders
-{
-    using AbstractBuilder.Examples.Entities;
-
-    internal class PointBuilder : RecordBuilder<Point>
-    {
-        public PointBuilder As2DX10Y20()
-        {
-            return Set<PointBuilder, double>(p => p.X, () => 10)
-                .Set<PointBuilder, double>(p => p.Y, () => 20);
-        }
-    }
-}
