--- UnitTests/Examples/Entities/BookCite.cs
@@ -1,4 +1,0 @@
-namespace AbstractBuilder.Examples.Entities
-{
-    internal record BookCite(string Title, string Author, string Publisher);
-}
