--- UnitTests/Examples/Entities/DatedBookCite.cs
@@ -1,7 +1,0 @@
-namespace AbstractBuilder.Examples.Entities
-{
-    internal record DatedBookCite(string Title = "Unknown", string Author = "Anonymous", string Publisher = null, int Year = 2022)
-        : BookCite(Title, Author, Publisher)
-    {
-    }
-}
