--- UnitTests/Examples/Entities/Point.cs
@@ -1,4 +1,0 @@
-namespace AbstractBuilder.Examples.Entities
-{
-    internal readonly record struct Point(double X, double Y, double Z);
-}
