--- UnitTests/Examples/Entities/WebBookCite.cs
@@ -1,4 +1,0 @@
-namespace AbstractBuilder.Examples.Entities
-{
-    internal record WebBookCite(BookCite Cite, string Url, int Year, char Initial, bool Free, double? Price);
-}
