+++ src/AbstractBuilder.sln
@@ -1,0 +1,35 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30002.166
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AbstractBuilder", "AbstractBuilder\AbstractBuilder.csproj", "{6FA7B93F-DBD4-4D31-9E72-B480A6DAA879}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UnitTests", "..\tests\UnitTests\UnitTests.csproj", "{4A817D41-93BD-4AFB-A6A5-8D2F71D5ABE6}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{D9E1A672-5DC0-49CB-B1EA-924DB32583D6}"
+	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6FA7B93F-DBD4-4D31-9E72-B480A6DAA879}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6FA7B93F-DBD4-4D31-9E72-B480A6DAA879}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6FA7B93F-DBD4-4D31-9E72-B480A6DAA879}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6FA7B93F-DBD4-4D31-9E72-B480A6DAA879}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4A817D41-93BD-4AFB-A6A5-8D2F71D5ABE6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4A817D41-93BD-4AFB-A6A5-8D2F71D5ABE6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4A817D41-93BD-4AFB-A6A5-8D2F71D5ABE6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4A817D41-93BD-4AFB-A6A5-8D2F71D5ABE6}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {7A342661-10E7-4A9E-9C14-DA312D1E6F7C}
+	EndGlobalSection
+EndGlobal
