+++ src/AbstractBuilder/AbstractBuilder.csproj
@@ -1,0 +1,32 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>netstandard2.0;netstandard2.1;net5.0;net6.0</TargetFrameworks>
+    <TargetFrameworks Condition="'$(OS)' == 'Windows_NT'">net45;netstandard2.0;netstandard2.1;net5.0;net6.0</TargetFrameworks>
+    <Authors>Pablo Caballero</Authors>
+    <Description>AbstractBuilder is a friendly object builder pattern for testing in .NET</Description>
+    <PackageLicenseFile>License.txt</PackageLicenseFile>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
+    <PackageProjectUrl>https://github.com/p-caballero/AbstractBuilder</PackageProjectUrl>
+    <RepositoryUrl>https://github.com/p-caballero/AbstractBuilder</RepositoryUrl>
+    <RepositoryType>GIT</RepositoryType>
+    <Copyright>Copyright (c) 2020-2023 Pablo Caballero</Copyright>
+    <AssemblyVersion>1.5.0.0</AssemblyVersion>
+    <FileVersion>1.5.0.0</FileVersion>
+    <Version>1.5.0</Version>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <None Remove=".editorconfig" />
+    <None Include="..\..\License.txt">
+      <Pack>True</Pack>
+      <PackagePath></PackagePath>
+    </None>
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Internal\" />
+  </ItemGroup>
+
+</Project>
