+++ tests/UnitTests/Examples/Builders/BuilderWithoutSeedCtor.cs
@@ -1,0 +1,12 @@
+namespace AbstractBuilder.Examples.Builders
+{
+    using AbstractBuilder.Examples.Entities;
+
+    internal class BuilderWithoutSeedCtor : AbstractBuilder<Car>
+    {
+        public BuilderWithoutSeedCtor()
+            : base(() => new Car())
+        {
+        }
+    }
+}
