+++ tests/UnitTests/Examples/Builders/DatedBookCiteBuilder.cs
@@ -1,0 +1,13 @@
+namespace AbstractBuilder.Examples.Builders
+{
+    using AbstractBuilder;
+    using AbstractBuilder.Examples.Entities;
+
+    internal class DatedBookCiteBuilder : RecordBuilder<DatedBookCite>
+    {
+        public DatedBookCiteBuilder WithYear()
+        {
+            return Set<DatedBookCiteBuilder, int>(x => x.Year, () => 1902);
+        }
+    }
+}
