+++ tests/UnitTests/Examples/Entities/BookCite.cs
@@ -1,0 +1,4 @@
+namespace AbstractBuilder.Examples.Entities
+{
+    internal record BookCite(string Title, string Author, string Publisher);
+}
