+++ tests/UnitTests/Examples/Entities/Car.cs
@@ -1,0 +1,19 @@
+namespace AbstractBuilder.Examples.Entities
+{
+    using System;
+
+    internal class Car
+    {
+        public const string DefaultModel = "UNKNOW MODEL";
+
+        public static readonly string DefaultColor = System.Drawing.Color.Black.Name;
+
+        public int Id { get; set; }
+
+        public string Model { get; set; } = DefaultModel;
+
+        public int NumDoors { get; set; }
+
+        public string Color { get; set; } = DefaultColor;
+    }
+}
