+++ tests/UnitTests/Examples/Entities/FerraryModels.cs
@@ -1,0 +1,9 @@
+namespace AbstractBuilder.Examples.Entities
+{
+    public static class FerraryModels
+    {
+        public const string Ferrari208Gts = "FERRARI 208 GTS";
+
+        public const string Ferrari308Gtbi = "FERRARI 308 GTBI";
+    }
+}
